添加了一些表

+++ database/migrations/20190328084352-init-users.js
@@ -4,7 +4,7 @@ module.exports = {
 	up: async (queryInterface, Sequelize) => {
 		const { INTEGER, DATE, STRING } = Sequelize;
 		await queryInterface.createTable("users", {
-			id: { type: INTEGER, primaryKey: true, autoIncrement: true },
+			id: { type: INTEGER(11), primaryKey: true, autoIncrement: true },
 			name: STRING(30),
 			age: INTEGER,
 			created_at: DATE,
@@ -16,3 +16,4 @@ module.exports = {
 		await queryInterface.dropTable("users");
 	}
 };
+
